Started valid-ready interfaces
Here we also have changes that make sure that
injected wires (due to named wired from tracer)
don't disturb the topology of the nets.

This is important to ensure that interfaces
have a straight back-pressure topology.

What still needs to happen is to intefaces
to *check* that the topology is supported.