Reworked Number simulation.
This is a good checkin point: the reworked Number simulation
(Number.SimValue) supports fractional types now, though it's
**COMPLETELY** untested at this point. However, the integer
version of it seems to be working and passing all existing tests.

This is also much slower as the previous implementation, but at
least more consistent, and optimization can wait. Now, all sim
operations are declared in the Number.SimValue class and all
other paths and operator overloads boil down to calling these
functions.

__invert__ is also fixed, as in removed: It's impossible to invert
a Number.SimValue since we don't know the length of it. An
'invert' method is provided for this, which takes an extra
argument. This should normally not pose a problem as Junctions
and XNets have this extra information. However if one works
hard enough, one can get into a situation that doesn't work.
A specific exception is raised in that case to inform the user.