Removing 'simulate' method for ExplicitAdaptor.
First of all, it doesn't work, and second, it's not needed:
they generated netlist in Body supports both RTL and sim.

BTW: the reason it doesn't work is that the netlist from body
drives signals that 'simulate' tried to override. That's not a nice
thing to do, plus it's impossible: we're attempting the drive the
middle of an XNet.