Added two new RV primitives: delay_line and pacer
'delay_line' has a fixed (minimum) latency, that is it takes
data at least N cycles to go through it. It's a hand-shaked
shift register in essence, though it's very similar to a FIFO in
implementation. The only difference is that it only lets data
drain, when it's full.

'pacer' is a rate-controller. It can accept a new data only every
N-th cycle. It doesn't buffer the data, only controls ready/valid
signals and even that is combinatorially passed through.

These two new primitives, although also synthesizable are
mostly helpful for developing high-level ready/valid models.

They allow for (more) cycle-accurate modelling of building blocks
where the abstract model is zero latency, but the actual implementation
would either be multi-cycle or longer latency (or a combination).

The two primitives pass simulation tests.