Fixed missing wire generation.
This was an instance where wire names could be emitted
without an assignment to it, if - later - a constant
assignment in the same scope appears.

Maybe there's a better way of dealing with this, but
it is at least correct now. The problem is in the order
of sub-module visitation, where the ConstantModule
gets visited during 'generate' later then the user.

Now, module visitation order is already optimized somewhat
so I'm not sure if there is an easy fix.