SelectOne allows multiple active selectors.
It is actually valid to have multiple active selectors
as long as the selected value is the same.

Since the RTL representation of the logic is an
AND-OR tree, such an input case is fine, so
the simulated behavior should match that as well.
Only assign 'None' to the output, if the values
mismatch.