Changed the way sim_value is reported for composites.
This is quite possibly blowing up in some TBs at the moment,
just not in the ones that are running in regression tests.

The change is that sim_value is now an object, containing members
for ... well, the members of the composite. So the normal '.'-notation
to access the value of individual members works.

The class for this is created on the fly and the members are 'dropped'
in as needed, recursively.

As a consequence, the RV checker API has changed as well.