Fixed FIFO buffer that got broken by the RAM conflict-resolution change.
When the RAM went from read-new-data to read-old-data if
the two ports conflicted on the address, the FIFO buffer logic
broke. I have a fix for it here, albeit not the best.

We probably want to re-visit this at some point, incorporate
the extra cycle of latency and make output_data registered.