NetType is now a type.
This is a rather fundamental change: all NetTypes are now proper types.
We also have NetTypeMeta, which is what all NetTypes are instances of.
Enum is renamed to EnumNet to remove the name collision.
Generic types are now a bit harder to make (generic Structs for example),
but there probably is some syntactic suggar that can be introduced
to make that easier.

At this point Nets and Junctions are not instances of NetTypes,
but eventually that'll come too.