Made Netlist the root object for Silicon.
This is instead of the 'top level' module that was the old way.

The problem of course is that the call-order on Netlist is now
not enforced: it's possible to try to 'generate' without 'elaborate'
for instance. But that's for a later day.

There's also a little confusion about how long a global Netlist
instance is needed. This should be cleaned up as well.