Added enable and clock-edge support to Reg.
There's now a GenericReg that supports
both Negative and Positive edges.
'Reg' is the version with Positive edges.
There's a NegReg and a PosReg variant too
which are not generic modules.

All of these support now the clock_en auto-input,
which, when connected creates a clock-enable.

Finally, auto-inputs now can also have default
values, though I'm not sure that's such a great idea.