Optional ports get assigned values in instantiation scope.
There is an important bugfix here, which get exposed by this work
but I think is unrelated:

If a sub-module loops back in input to its output, in the enclosing scope
we could have a local driver (such as a ConstModule). In that case the
XNet will have the driver correctly identified as the ConstModule,
but during RTL generation we would create an output binding as well
for the sub-module, creating two drivers for the same net.

The solution in this change is to skip generating output bindings for
submodules under these circumstances, though it's not all that easy
to correctly identify 'these circumstances'. It's a bit hacky and I'm not
100% sure it works for all cases.

Also added here is the 'can_be_deleted' parameter for ports. If
specified, this allows ports to be deleted if they are left unassigned,
even if they are not AutoInput instances.
This is used on _SelectOneHot.default_port. Probably not that
useful for anything but some primitives.