Fixed (hopefully) loopback problems.
This is a large change that introduces TrivialAdaptors.
These are used to break an XNet into pieces at
critical points to ensure that if a net re-enters the
same scope two or more times (due to internal or
external loopbacks), each section has its own driver
within that scope.

There are two additional changes in here:
- made port._sinks private (and have access methods)
- fixed PhiSlice insertion where _sinks of _partial_sources
  were not properly cleaned up