Fixed up netlist manipulation for local wires.
This change removes local wire splicing into the netlist.
Instead, it adds the local names to the objects (or rather
registers the names in the symbol table for the objects)
they refer to. This then had a lot of ripple-effects.

Most importantly, local wires are registered as soft-symbols
as they can interfere with attribute-symbols, so they
might need renaming. However, soft-symbols are not
used as the primary name for RTL generation and
not even generated assignments for if not used.

Second, some objects, such as UniSlicers are not
resolved at the moment (that is, they are JunctionBases
and not Junctions) so they don't appear in XNets and
as a consequence their local names are not used.

This is good for ScopedPorts (we don't want those) but
not so much for UniSlicers.

It's also not so good that input and output ports which
get assigned to local wires are still referred to as their
port names, those being used/assigned at the top
of RTL generation.

Now, none of this is a correctness issue, but they do
degrade RTL readability.

Other important details for code-changes:
- register_local_wire is gone
- Simulator supports unconnected outputs properly now.
- SymbolTable now returns OrderedSets instead of frozensets.