Worked around pydevd debugger in tracer.
This is something that either VSCode started doing
recently or pydevd changed recently, but either way:

The pydev debugger is inserting a bunch of function
calls into the function-call mechanism, probably
at the bytecode level. This means that we mis-attribute
the enable stack that we maintain and essentially
tracing is disabled, when run under debugger.

The work-around is to special-case frames under 'pydevd'
and bail on the tracer. Hopefully this trick will work with
other debuggers, such as pycharm as well, but of course
who knows...