Simulator now tracks changes better.
We now properly track value changes
and only fire events if a wire gets assigned
a *different* value then what it currently
has. This cuts down on the number of
events fired and also fixes spurious
edge-detections.

There's also a change here that exposes
when a simulation value was last changed
on a port: the 'last_changed' property.