RV interface broke with the recent net-type changes.
This is a curious one: with net-type being part of the
class for a net instance, rv_interface didn't properly
cache data_member_type. Instead, it returned a new
(identical) type for each port instance. This of course
led to trouble if we wanted to anything with these
types, such as a Select.

The fix is to get to this method always through 'get_net_type()'
which fixes the class hierarchy.

Also, set_data_members is broken a little: we normally
don't know output net types, when this method is called.
As a consequence, it can't really set up member assignments.

For now, there's only a check and a syntax error for this, but
the right thing to do would be to register a call-back for when
the net-type does get known and do the assignment there.