Edge handling in simulation fixed.
This is an important change: we reset edges in each loop
through the same 'now' timestamp to make sure we don't
keep re-interpreting edges over and over again.

TODO: this should be better optimized, essentially treating
epsilon time better. For now, it seems to work, but probably
is rather slow.