Started BasicMem.
This would be a primitive async mem for simulation only.
It should handle any number of ports with any sizes.

It passes it's first simulation, but more is needed.

It should be the basis for more complex memories, but
only for simulation, I think. Since the registers are supposed
to be generated into a neat little bundle of RTL for the tools
to recognize and infer the memory, even if the sim model
is the same as the generated RTL, we can't use the sim model
(hierarchy really) to generate said RTL.