Verilog is crazy!
First off, it has this crazy habit of determining operand size
by looking at the destination type of the *whole* expression,
not just the input arguments. This is documented in

http://ebook.pldworld.com/_eBook/FPGA%EF%BC%8FHDL/-Eng-/IEEE%201364-1995%20standard.%20Verilog%20hardware%20description%20language.pdf

chapter 4.4

The work-around suggested in the standard is to 'add a proper-length  0'
to the expression to force evaluation at the expect length.

Another way is to break out the sub-expression into its own expression
and assign it to a properly sized variable.

This change implements both, though the detection were it's needed
is a bit overly conservative, in that it looks at the sub-expression
and not the full one. It 'adds 0' if there is no suitable name in the source
for the sub-expression, otherwise it breaks it out.

The second change has to do with the fact that original Verilog '95
didn't have any way of casting the expression size. Or rather, the only
way to do that is - again - by breaking it out into a separate variable
assignment.

This feature is not well tested at the moment and is certainly not
working for enums, but the changes are in here.