Simulator now has asserts.
This is in preparation for exporting test-benches.
The idea is that we will have the ability to export
stimulus from 'simulate' methods that are marked
somehow as well as generate Verilog asserts from
Simulator.sim_assert calls.

This hopefully will allow to re-run the test-benches
in Verilog and expect the same results as in Python.