RvArbiter passes simulation test.
It also has now a FIFO interface intrnally
to store outstanding response. HOWEVER!!!!

This FIFO has to have a depth of 1, otherwise
there is a combinational loop.

The problem is that FIFO output data
depends on the (FIFO-internal) bypass mux.
This in turn depends on output.ready,
which in turn depends on which xxx_response
port we select, which is of course
dependent on the output data from
the FIFO as that determines which
port to listen to.

I think the FIFO will need yet more work...