Testing dual-port memory behavior.
There is an error in Silicon here:
A simple dual-port memory with
'read-new-data' behavior should
read new data *across* ports as well.
That is, if I write address X with data Y
on port 1 and read address X on the other
in the same cycle, I should get Y, not
whatever was in there before.

At least that's what iVerilog tells me
and that's not what Silicon tells me.

Gaaahh!!!!