FiFo output mux doesn't depend on read/valid.
The output data from the FIFO contains a bypass mux
to help with the read-old-data behavior of the underlying
memory. The selector for this mux should not depend
on the immediate state of read/valid signals on the output
interface because that can cause insidious combinational
loops (not to mention degraded speed).