Work towards separating out arbitration logic.
This still needs a lot of work, but at least something starts to take
shape. I still don't know how to best inject the arbiter logic into
an actual RVArbiter. I'm also unsure of how to best expose the
right state of RVArbiter to the logic, things such as burst-handling.